refactor(test): reorganize test-files - move to test/hooks - move to test/methods - rename test/helpers to test/test-helpers
