+++ package.json
@@ -77,6 +77,7 @@
     "@typescript-eslint/parser": "^4.31.0",
     "chai": "^4.3.4",
     "cross-env": "^7.0.3",
+    "date-fns": "^2.23.0",
     "eslint": "^7.32.0",
     "eslint-config-standard-with-typescript": "^21.0.1",
     "feathers-memory": "^4.1.0",
