+++ test/methods/check-unique.test.ts
@@ -1,9 +1,9 @@
 import { assert } from 'chai';
 import feathers, { Application } from '@feathersjs/feathers';
 import feathersMemory, { Service } from 'feathers-memory';
-import authLocalMgnt from '../src/index';
-import { timeoutEachTest } from './helpers/config';
-import { AuthenticationManagementService } from '../src/services';
+import authLocalMgnt from '../../src/index';
+import { timeoutEachTest } from '../test-helpers/config';
+import { AuthenticationManagementService } from '../../src/services';
 
 const makeUsersService = options =>
   function (app: Application) {
