+++ test/methods/resend-verify-signup.test.ts
@@ -1,15 +1,15 @@
 import { assert } from 'chai';
 import feathers, { Application } from '@feathersjs/feathers';
 import feathersMemory, { Service } from 'feathers-memory';
-import authLocalMgnt from '../src/index';
+import authLocalMgnt from '../../src/index';
 import {
   SpyOn,
   aboutEqualDateTime,
   makeDateTime
-} from './helpers';
-import { timeoutEachTest, defaultVerifyDelay } from './helpers/config';
-import { UserTestDB, UserTestLocal } from './helpers/types';
-import { AuthenticationManagementService } from '../src/services';
+} from '../test-helpers';
+import { timeoutEachTest } from '../test-helpers/config';
+import { UserTestDB, UserTestLocal } from '../test-helpers/types';
+import { AuthenticationManagementService } from '../../src/services';
 
 const now = Date.now();
